pyproject.toml
[tool.poetry]
name = "catalogue"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
gunicorn = "^21.2.0"
pydantic = { version = "^1", extras=["dotenv"] }
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
alembic = "^1.11.1"
asyncpg = {version = "^0.28.0", extras = ["sa"]}
redis = {version = "^4.6.0", extras = ["hiredis"]}
httptools = "^0.6.0"
strawberry-graphql = { version = "^0.194.4", extras = ["fastapi"] }
prometheus-client = "^0.17.0"
prometheus-fastapi-instrumentator = "6.0.0"
opentelemetry-api = "^1.18.0"
opentelemetry-sdk = "^1.18.0"
opentelemetry-exporter-otlp = "^1.18.0"
opentelemetry-instrumentation = "^0.39b0"
opentelemetry-instrumentation-fastapi = "^0.39b0"
opentelemetry-instrumentation-redis = "^0.39b0"
opentelemetry-instrumentation-sqlalchemy = "^0.39b0"
loguru = "^0.7.0"
aiokafka = "^0.8.1"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["catalogue",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
plugins = ["strawberry.ext.mypy_plugin"]

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "CATALOGUE_ENVIRONMENT=pytest",
    "CATALOGUE_DB_BASE=catalogue_test",
]

[fastapi-template.options]
project_name = "catalogue"
api_type = "graphql"
enable_redis = "True"
enable_rmq = "None"
ci_type = "github"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "True"
kube_name = "catalogue"
enable_routers = "None"
enable_kafka = "True"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "None"
orm = "sqlalchemy"
self_hosted_swagger = "None"
prometheus_enabled = "True"
sentry_enabled = "None"
otlp_enabled = "True"
pydanticv1 = "True"
gunicorn = "True"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
